% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google_fonts.R
\name{activate_google_font}
\alias{activate_google_font}
\title{Use Google Font}
\usage{
activate_google_font(
  name,
  regular.wt = 400,
  bold.wt = 700,
  repo = "http://fonts.gstatic.com/",
  db_cache = TRUE,
  handle = curl::new_handle()
)
}
\arguments{
\item{name}{name of the font that will be searched in Google Fonts}

\item{regular.wt}{font weight for the regular font face, usually 400}

\item{bold.wt}{font weight for the bold font face, usually 700}

\item{repo}{the site that hosts the font files. Default is the official
repository \code{http://fonts.gstatic.com/} provided by
Google Fonts.}

\item{db_cache}{whether to obtain font metadata from a cache site. Using cache
is typically faster, but not as update-to-date as using the official
API. If \code{db_cache} is set to \code{FALSE}, then metadata
are retrieved from the Google Fonts API.}

\item{handle}{a curl handle object passed to \code{curl::curl_download()}.}
}
\description{
Adds a Google Font from \url{https://fonts.google.com/} and activate the showtext package. This function is particularly useful for using Google Fonts for \code{\link{[createLogo]createLogo}}.
}
\seealso{
\code{\link[curl]{handle}}
\code{\link[sysfonts]{font_add_google}}
\code{\link[showtext]{showtext_auto}}
}
