% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_roxygen2.R
\name{add_deprecated_call}
\alias{add_deprecated_call}
\title{Add \code{.Deprecated()} to Text}
\usage{
add_deprecated_call(text, new = NULL)
}
\arguments{
\item{text}{Character vector containing function declarations.}

\item{new}{(Optional) Argument passed to the new call to \code{.Deprecated()}.}
}
\description{
\code{.Deprecated()} is added after each function declaration in the text. Note that it does not check for an existing \code{.Deprecated()}.
}
\seealso{
\code{\link[stringr]{str_replace}}

Other modify roxygen2 documentation functions: 
\code{\link{add_deprecated_to_desc}()}
}
\concept{modify roxygen2 documentation functions}
