% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_roxygen2.R
\name{add_deprecated_to_desc}
\alias{add_deprecated_to_desc}
\title{Add "(Deprecated)" after the \verb{@description} Roxygen2 Tag}
\usage{
add_deprecated_to_desc(text)
}
\arguments{
\item{text}{Character vector containing function declarations.}
}
\description{
Add "(Deprecated)" after the \verb{@description} Roxygen2 Tag
}
\examples{
library(tidyverse)

# Adds "Deprecated" after each Roxygen2 Tag
add_deprecated_to_desc(text = "#' This is the title\n#' @description\n#' This is the description.")
}
\seealso{
\code{\link[stringr]{str_replace}}

Other modify roxygen2 documentation functions: 
\code{\link{add_deprecated_call}()}
}
\concept{modify roxygen2 documentation functions}
