% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{assign_colors}
\alias{assign_colors}
\title{FUNCTION_TITLE}
\usage{
assign_colors(
  x,
  taken_colors,
  shade_level = 2,
  seed = 100,
  exclude_color_regex
)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{taken_colors}{PARAM_DESCRIPTION}

\item{shade_level}{PARAM_DESCRIPTION, Default: 2}

\item{seed}{PARAM_DESCRIPTION, Default: 100}

\item{exclude_color_regex}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}
}
