% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{assign_colors_by}
\alias{assign_colors_by}
\title{FUNCTION_TITLE}
\usage{
assign_colors_by(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[colorspace]{hcl_palettes}}
}
