% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_labels.R
\name{authorize_google}
\alias{authorize_google}
\title{Authorize Google API Access}
\usage{
authorize_google()
}
\description{
Authorize Google API Access
}
\seealso{
Other Google: 
\code{\link{get_gmail_inbox}()},
\code{\link{get_gmail_labels}()},
\code{\link{get_gmail_messages_by_label}()},
\code{\link{get_unread_gmail}()},
\code{\link{get_unread_unlabelled_emails}()},
\code{\link{process_gmail_filters_xml}()}
}
\concept{Google}
