% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{auto_map_random_colors}
\alias{auto_map_random_colors}
\title{Randomly Select Colors}
\usage{
auto_map_random_colors(
  x,
  taken_colors,
  shade_level = 2,
  seed = 100,
  exclude_color_regex = "black|white|ivory|snow|beige|gray|grey"
)
}
\arguments{
\item{x}{Vector to map.}

\item{taken_colors}{(Optional) Colors to exclude from selection because they have
already been taken.}

\item{shade_level}{Shade level from 1 to 4, Default: 2}

\item{seed}{Default: 100}

\item{exclude_color_regex}{(Optional) Regex to exclude a subset of colors.}
}
\description{
Randomly Select Colors
}
\seealso{
\code{\link[dplyr]{filter}}
}
