% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{auto_map_to_palette}
\alias{auto_map_to_palette}
\title{Use Colorspace Palettes for Map Assignment}
\usage{
auto_map_to_palette(x, colorspace_palette = "diverge_hcl", ...)
}
\arguments{
\item{x}{Vector to map. All vectors are converted to character
before mapping.}
}
\description{
Instead of having to manually determine a map assignment,
a colorspace palette can be used to auto-assign to a given
vector.
}
\examples{
\dontrun{
if(interactive()){
  auto_map_to_color(x = as.character(x)) \%>\%
  view_colors()
 }
}
}
\seealso{
\code{\link[colorspace]{hcl_palettes}}
}
