% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{base_colors}
\alias{base_colors}
\title{FUNCTION_TITLE}
\usage{
base_colors()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}}
\code{\link[rubix]{filter_at_grepl}}
}
