% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_cache.R
\name{cache_if_interactive}
\alias{cache_if_interactive}
\title{Cache Only In Interactive}
\usage{
cache_if_interactive(object, dirs, overwrite = FALSE)
}
\arguments{
\item{object}{Object to be cached.}

\item{dirs}{Required. Path to the subdirectory within R.cache root path to
save the cache to.}
}
\description{
If called in an interactive session, the object will be cached.
}
\details{
This function is useful when reporting on a process that has already been
executed and does not require re-execution during the knitting process.
For example, a Neo4j database can be written in an interactive session at
which point the metadata for reporting purposes can be logged. Instead of forcing
the database to be instantiated when the R markdown is knit, knitr will
only pull up the cached log.
}
\seealso{
\code{\link[R.cache]{saveCache}},\code{\link[R.cache]{findCache}},\code{\link[R.cache]{loadCache}}
}
