% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_cache.R
\name{cache_object}
\alias{cache_object}
\title{Cache an Object}
\usage{
cache_object(object, dirs, overwrite = FALSE)
}
\arguments{
\item{object}{Object to be cached.}

\item{dirs}{Required. Path to the subdirectory within R.cache root path to
save the cache to.}
}
\description{
Cache an Object
}
\seealso{
\code{\link[R.cache]{saveCache}},\code{\link[R.cache]{findCache}},\code{\link[R.cache]{loadCache}}
}
