% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_time_remaining.R
\name{calculate_time_remaining}
\alias{calculate_time_remaining}
\title{Calculate Time Remaining}
\usage{
calculate_time_remaining(
  iteration,
  total_iterations,
  time_value_per_iteration,
  time_unit_per_iteration = c("seconds", "minutes", "hours", "days", "weeks", "months",
    "years", "milliseconds", "microseconds", "nanoseconds", "picoseconds")
)
}
\arguments{
\item{iteration}{Integer of the iteration.}

\item{total_iterations}{Integer of total iterations in the loop.}

\item{time_value_per_iteration}{Numeric of the time per iteration.}

\item{time_unit_per_iteration}{Time unit that is passed to \code{\link[lubridate]{duration}}. Options: c("seconds", "minutes", "hours", "days", "weeks", "months", "years",
"milliseconds", "microseconds", "nanoseconds", "picoseconds")}
}
\value{
Duration class object
}
\description{
Calculate the time remaining after an iteration of a
loop is completed.
}
\details{
DETAILS
}
