% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_diagram.R
\name{cat_custom_diagram}
\alias{cat_custom_diagram}
\title{Concatenate a Customized Diagram}
\usage{
cat_custom_diagram(
  flow,
  ...,
  border_style = "single",
  padding = 1,
  margin = 0,
  float = c("left", "center", "right"),
  col = NULL,
  background_col = NULL,
  border_col = col,
  align = c("left", "center", "right"),
  width = cli::console_width()
)
}
\arguments{
\item{flow}{String that alternates between a box label and arrows beginning
with either \code{-} or \code{<}.}

\item{...}{A list named by the string representing a node in the
\code{flow} string that contains a list of argument-value pair that will
be passed to the \code{cli::cat_boxx} function.}
}
\description{
More custom features passed onto \code{cli::cat_boxx}
by each node in a diagram.
}
\details{
Example of a named list of customizations passed to the function:\preformatted{flow <-
  "Test Custom Box -> Test Custom Box 2 -> Test Custom Box 3"

 cat_custom_diagram(
   flow = flow,
   'Test Custom Box 2' =
      list(header = "Database"),
   'Test Custom Box 3' =
      list(border_style = "single-double"))
}
}
\seealso{
\code{\link[cli]{console_width}},\code{\link[cli]{cat_line}}
\code{\link[rlang]{list2}},\code{\link[rlang]{parse_expr}}
\code{\link[purrr]{transpose}},\code{\link[purrr]{map}}
\code{\link[utils]{capture.output}}
\code{\link[glue]{glue}}
}
