% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_diagram.R
\name{cat_diagram}
\alias{cat_diagram}
\title{Concatenate a Diagram}
\usage{
cat_diagram(
  flow,
  border_style = "single",
  padding = 1,
  margin = 0,
  float = c("left", "center", "right"),
  col = NULL,
  background_col = NULL,
  border_col = col,
  align = c("left", "center", "right"),
  width = cli::console_width()
)
}
\arguments{
\item{flow}{String that alternates between a box label and arrows beginning
with either \code{-} or \code{<}.}
}
\description{
Concatenate a diagram in a left to right direction.
}
\seealso{
\code{\link[cli]{console_width}},\code{\link[cli]{cat_line}}
}
