% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_html_color.R
\name{cat_html_color}
\alias{cat_html_color}
\title{FUNCTION_TITLE}
\usage{
cat_html_color(text, color)
}
\arguments{
\item{text}{PARAM_DESCRIPTION}

\item{color}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
}
