% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{cat_html_link}
\alias{cat_html_link}
\title{Make a HTML link to a new tab}
\usage{
cat_html_link(link, label)
}
\arguments{
\item{link}{HTML link}

\item{label}{Label for HTML link}
}
\description{
Make a HTML link to a new tab
}
