% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_html.R
\name{cat_link_to_report}
\alias{cat_link_to_report}
\title{Add Link Referencing Another Report}
\usage{
cat_link_to_report(rmd_title)
}
\description{
HTML link to another report based by
title. Note that this function does not
check if the link is valid and that the corresponding
source file exists.
}
\seealso{
\code{\link[xfun]{file_ext}}
\code{\link[stringr]{str_replace}}
}
