% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color_to_hex}
\alias{color_to_hex}
\title{Color to Hex}
\usage{
color_to_hex(x, alpha = FALSE)
}
\arguments{
\item{x}{Color name as string.}

\item{alpha}{logical value indicating whether the alpha channel (opacity)
    values should be returned.}
}
\value{
Vector of length 1 or greater of hex codes matched to the RGB derived
from the color name.
}
\description{
Get the hex code for a color given by its name.
}
\examples{
color_to_hex("blue")
}
\seealso{
\code{\link[grDevices]{col2rgb}},\code{\link[grDevices]{rgb}}
}
