% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{colors_hex_map}
\alias{colors_hex_map}
\title{Map Between Base R Colors and Hex Codes}
\usage{
colors_hex_map()
}
\value{
Vector of hexadecimal codes named by the source
color name.
}
\description{
All colors returned by base::colors() are
mapped to their corresponding hex codes.
}
