% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pkg.R
\name{create_data_repo_dirs}
\alias{create_data_repo_dirs}
\title{Create Repo Directories}
\usage{
create_data_repo_dirs(path = getwd())
}
\arguments{
\item{path}{Path to the package root that contains the \verb{data-raw/}, \verb{data/}, and \verb{R/} folders. `}
}
\description{
Creates the \code{data-raw}, \code{data}, and \code{R} directories within the \code{path} if they do not already exist.
}
