% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medallion.R
\name{create_favicons}
\alias{create_favicons}
\title{Create and Move Favicons}
\usage{
create_favicons(overwrite = FALSE)
}
\description{
This package first runs the build_favicons() function in the pkgdown package,
which auto-detects a logo.svg or logo.png file and runs it through an API.
The function makes a copy of the favicon output to the "man/figures"
directory for use.
}
