% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medallion.R
\name{create_medallion}
\alias{create_medallion}
\title{Create Medallion}
\usage{
create_medallion(
  sourceImg,
  package,
  white_around_sticker = TRUE,
  s_x = 0.8,
  s_y = 0.75,
  s_width = 0.4,
  s_height = 0.5,
  p_x = 1,
  p_y = 1.4,
  p_color = "#FFFFFF",
  p_family = "Aller_Rg",
  p_size = 8,
  h_size = 1.2,
  h_fill = "#1881C2",
  h_color = "#87B13F",
  spotlight = FALSE,
  l_x = 1,
  l_y = 0.5,
  l_width = 3,
  l_height = 3,
  l_alpha = 0.4,
  url = "",
  u_x = 1,
  u_y = 0.08,
  u_color = "black",
  u_family = "Aller_Rg",
  u_size = 1.5,
  u_angle = 30,
  filename = paste0(package, ".png"),
  asp = 1,
  dpi = 300
)
}
\arguments{
\item{package}{package name}

\item{white_around_sticker}{default to FALSE. If set to TRUE, it puts
white triangles in the corners}

\item{s_x}{x position for subplot}

\item{s_y}{y position for subplot}

\item{s_width}{width for subplot}

\item{s_height}{height for subplot}

\item{p_x}{x position for package name}

\item{p_y}{y position for package name}

\item{p_color}{color for package name}

\item{p_family}{font family for package name}

\item{p_size}{font size for package name}

\item{h_size}{size for hexagon border}

\item{h_fill}{color to fill hexagon}

\item{h_color}{color for hexagon border}

\item{spotlight}{whether add spotlight}

\item{l_x}{x position for spotlight}

\item{l_y}{y position for spotlight}

\item{l_width}{width for spotlight}

\item{l_height}{height for spotlight}

\item{l_alpha}{maximum alpha for spotlight}

\item{url}{url at lower border}

\item{u_x}{x position for url}

\item{u_y}{y position for url}

\item{u_color}{color for url}

\item{u_family}{font family for url}

\item{u_size}{text size for url}

\item{u_angle}{angle for url}

\item{filename}{filename to save sticker}

\item{asp}{aspect ratio, only works if subplot is an image file}

\item{dpi}{plot resolution}
}
\description{
Create Medallion
}
