% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pkg.R
\name{data_package_helpers}
\alias{data_package_helpers}
\title{Data Package Helpers}
\arguments{
\item{repo_path}{Path to a local clone of the target repo.}

\item{conn}{Connection object.}
}
\description{
Data Package Helpers are functions that support the maintenance of the data that is exported in packages.
}
