% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare.R
\name{declare_args}
\alias{declare_args}
\title{Create R Objects from Formal Argument Default Values}
\usage{
declare_args(fun)
}
\description{
Declare formal arguments of a function along with their default values as R objects
}
\concept{declaration functions}
