% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declare.R
\name{declareArgs}
\alias{declareArgs}
\title{Create R Objects from Formal Argument Default Values}
\usage{
declareArgs(fun)
}
\description{
Declare formal arguments of a function along with their default values as R objects
}
\concept{declaration functions}
