% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyBakeOven-package.R
\docType{package}
\name{easyBakeOven-package}
\alias{easyBakeOven}
\alias{easyBakeOven-package}
\title{easyBakeOven: Helpers for Package Authoring}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

R Package that facilitates various aspects of R
    package authorship and deployment in tandem with the usethis,
    devtools, and sinew packages. Features include generating pkgdown yaml
    skeletons as well as providing customizability to the roxygen2
    skeletons made by the sinew package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://meerapatelmd.github.io/easyBakeOven/}
  \item \url{https://github.com/meerapatelmd/easyBakeOven/}
  \item Report bugs at \url{https://github.com/meerapatelmd/easyBakeOven/issues/}
}

}
\author{
\strong{Maintainer}: Meera Patel, MD \email{patelmeeray@gmail.com} (\href{https://orcid.org/0000-0002-3401-3948}{ORCID})

}
\keyword{internal}
