% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_labels.R
\name{get_gmail_inbox}
\alias{get_gmail_inbox}
\title{Get Inbox}
\usage{
get_gmail_inbox(num_results = 1000)
}
\description{
Get Inbox
}
\seealso{
Other Google: 
\code{\link{authorize_google}()},
\code{\link{get_gmail_labels}()},
\code{\link{get_gmail_messages_by_label}()},
\code{\link{get_unread_gmail}()},
\code{\link{get_unread_unlabelled_emails}()},
\code{\link{process_gmail_filters_xml}()}
}
\concept{Google}
