% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_labels.R
\name{get_gmail_messages_by_label}
\alias{get_gmail_messages_by_label}
\title{Get Gmail Messages By Label}
\usage{
get_gmail_messages_by_label(
  search = NULL,
  num_results = 10,
  include_spam_trash = FALSE
)
}
\arguments{
\item{search}{query to use, same format as gmail search box.}

\item{num_results}{the number of results to return.}

\item{include_spam_trash}{boolean whether to include the spam and trash folders in the search}
}
\description{
Get Gmail Messages By Label
}
\seealso{
Other Google: 
\code{\link{authorize_google}()},
\code{\link{get_gmail_inbox}()},
\code{\link{get_gmail_labels}()},
\code{\link{get_unread_gmail}()},
\code{\link{get_unread_unlabelled_emails}()},
\code{\link{process_gmail_filters_xml}()}
}
\concept{Google}
