% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_labels.R
\name{get_unread_gmail}
\alias{get_unread_gmail}
\title{Get Unread Emails}
\usage{
get_unread_gmail(
  days_back = 90,
  num_results = 1000,
  include_spam_trash = FALSE
)
}
\description{
Get Unread Emails
}
\seealso{
Other Google: 
\code{\link{authorize_google}()},
\code{\link{get_gmail_inbox}()},
\code{\link{get_gmail_labels}()},
\code{\link{get_gmail_messages_by_label}()},
\code{\link{get_unread_unlabelled_emails}()},
\code{\link{process_gmail_filters_xml}()}
}
\concept{Google}
