% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{hex_colors_map}
\alias{hex_colors_map}
\title{FUNCTION_TITLE}
\usage{
hex_colors_map(x)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
