% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{hex_to_color}
\alias{hex_to_color}
\title{FUNCTION_TITLE}
\usage{
hex_to_color(hex)
}
\arguments{
\item{hex}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
