% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_report_vars.R
\name{list_all_variables}
\alias{list_all_variables}
\title{FUNCTION_TITLE}
\usage{
list_all_variables(file)
}
\arguments{
\item{file}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[stringr]{str_replace}}
}
