% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-list_report_vars.R
\name{list_missing_variables}
\alias{list_missing_variables}
\title{FUNCTION_TITLE}
\usage{
list_missing_variables(file)
}
\arguments{
\item{file}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[glue]{glue}}
\code{\link[stringr]{str_replace}}
}
