% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{list_template_variables}
\alias{list_template_variables}
\title{List all the Bracketed Variables}
\usage{
list_template_variables(template_path)
}
\arguments{
\item{template_path}{Path to the template file to use. It can point to a
template Rmd in your local machine.}
}
\description{
List all the Bracketed Variables
}
\seealso{
\code{\link[xfun]{read_utf8}}
}
