% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_cache.R
\name{load_cached_object}
\alias{load_cached_object}
\title{Load Cached Object}
\usage{
load_cached_object(object, dirs)
}
\arguments{
\item{object}{Object that was cached.}

\item{dirs}{Required. Path to the subdirectory within R.cache root path to
save the cache to.}
}
\description{
Load Cached Object
}
\seealso{
\code{\link[R.cache]{saveCache}},\code{\link[R.cache]{findCache}},\code{\link[R.cache]{loadCache}}
}
