% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_arg_declaration}
\alias{make_arg_declaration}
\title{Make Arg Declaration}
\usage{
make_arg_declaration(fun)
}
\description{
To automatically assign argument objects with the default value, see \code{\link{declareArgs}}
}
\examples{
make_default_args(read.csv)
make_internal_args(read.csv)
make_args(read.csv)
make_arg_declaration(read.csv)
}
\seealso{
Other make functions: 
\code{\link{make_args}()},
\code{\link{make_description_links}()},
\code{\link{make_description}()},
\code{\link{make_imports}()}
}
\concept{make functions}
