% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_args}
\alias{make_args}
\title{Make Args}
\usage{
make_args(fun)
}
\description{
Make Args
}
\examples{
make_default_args(read.csv)
make_internal_args(read.csv)
make_args(read.csv)
make_arg_declaration(read.csv)
}
\seealso{
Other make functions: 
\code{\link{make_arg_declaration}()},
\code{\link{make_description_links}()},
\code{\link{make_description}()},
\code{\link{make_imports}()}
}
\concept{make functions}
