% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_description_links}
\alias{make_description_links}
\title{Make URLS For DESCRIPTION}
\usage{
make_description_links(github_user, repo)
}
\description{
Make the links to the Repo, GitHub Pages, and Issues Pages. To make these links in addition to the entire DESCRIPTION contents, see \code{\link{make_description}}.
}
\seealso{
Other make functions: 
\code{\link{make_arg_declaration}()},
\code{\link{make_args}()},
\code{\link{make_description}()},
\code{\link{make_imports}()}

Other DESCRIPTION functions: 
\code{\link{make_description}()},
\code{\link{make_imports}()},
\code{\link{read_description}()}
}
\concept{DESCRIPTION functions}
\concept{make functions}
