% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_imports}
\alias{make_imports}
\title{Make DESCRIPTION Imports from NAMESPACE}
\usage{
make_imports()
}
\description{
Make DESCRIPTION Imports from NAMESPACE
}
\seealso{
Other make functions: 
\code{\link{make_arg_declaration}()},
\code{\link{make_args}()},
\code{\link{make_description_links}()},
\code{\link{make_description}()}

Other DESCRIPTION functions: 
\code{\link{make_description_links}()},
\code{\link{make_description}()},
\code{\link{read_description}()}
}
\concept{DESCRIPTION functions}
\concept{make functions}
