% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_new_package.R
\name{make_new_package}
\alias{make_new_package}
\title{Start a New Package}
\usage{
make_new_package(path, open = FALSE, spellcheck_vignettes = TRUE)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}

\item{spellcheck_vignettes}{Spell check is automatically enabled for the new package, but if TRUE, will also spell check all rmd and rnw files in the vignettes/ folder.}
}
\description{
Starts a new package using a designated path to a local directory, where the folder is created, README.md and NEWS.md files are added, git is initialized and linked to a remote of the same name.
}
\seealso{
\code{\link[usethis]{create_package}},\code{\link[usethis]{proj_activate}},\code{\link[usethis]{use_readme_rmd}},\code{\link[usethis]{use_news_md}},\code{\link[usethis]{use_git}}
}
