% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.R
\name{make_pkgdown_yaml}
\alias{make_pkgdown_yaml}
\title{Make Pkgdown YAML Template}
\usage{
make_pkgdown_yaml()
}
\description{
Prints a pkgdown yaml template to the R console for copy-and-pasting. To get true references to copy-and-paste into the pkgdown yaml, see \code{\link{make_pkgdown_yaml_refs}}.
}
