% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.R
\name{make_pkgdown_yaml_refs}
\alias{make_pkgdown_yaml_refs}
\title{Make a List of References for Pkgdown YAML}
\usage{
make_pkgdown_yaml_refs()
}
\description{
Prints yaml format references for copy-and-pasting into the pkgdown yaml file. To print a template, see \code{\link{make_pkgdown_yaml}}.
}
\seealso{
\code{\link[stringr]{str_remove}}
}
