% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maker.R
\name{make_tribble}
\alias{make_tribble}
\title{Make Tribble Expression}
\usage{
make_tribble(data, quote = "'")
}
\arguments{
\item{data}{Data to convert to expression.}

\item{quote}{How strings will be flanked, Default: '''}
}
\description{
Make Tribble Expression
}
