% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vignette.R
\name{make_vignette_asis}
\alias{make_vignette_asis}
\title{Add a Static Vignette with R.rsp}
\usage{
make_vignette_asis(file, vignette_title, package = basename(getwd()))
}
\arguments{
\item{file}{Name of the pdf or html file in the \verb{vignettes/} directory. The full
path is not required.}

\item{vignette_title}{The title of the vignette that will be indexed.}

\item{package}{Package that the vignette belongs to. Defaults to the basename
of the current working directory.}
}
\description{
Make the asis file for static PDF or HTML vignettes which is named the same as
the \code{file} argument with".asis" appended. The asis file will not be written
if it already exists, if the \code{file} file argument does not exist in the
\code{vignettes} folder, or if \code{file} does not have a '.pdf' or '.html' extension.

Compared to \code{\link{make_vignette_rnw}}, make_vignette_asis allows using
static HTML vignettes. However, it requires adding R.rsp to the Suggests and
VignetteBuilder fields in the DESCRIPTION file to execute.
}
\seealso{
Other make vignette functions: 
\code{\link{make_vignette_rnw}()}
}
\concept{make vignette functions}
