% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vignette.R
\name{make_vignette_rnw}
\alias{make_vignette_rnw}
\title{Make Rnw}
\usage{
make_vignette_rnw(pdf_file, vignette_title, package = basename(getwd()))
}
\arguments{
\item{pdf_file}{Name of the pdf file in the \verb{vignettes/} directory. The full
path is not required.}

\item{vignette_title}{The title of the vignette that will be indexed.}

\item{package}{Package that the vignette belongs to. Defaults to the basename
of the current working directory.}
}
\description{
Make the Sweave Rnw file for static PDF vignettes with the same file name.
The Rnw file will not be written if it already exists, if the \code{pdf_file}
file argument does not exist in the \code{vignettes} folder, or if \code{pdf_file} does
not have a '.pdf' extension.
}
\seealso{
Other make vignette functions: 
\code{\link{make_vignette_asis}()}
}
\concept{make vignette functions}
