% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeArgDeclaration}
\alias{makeArgDeclaration}
\title{Make Arg Declaration}
\usage{
makeArgDeclaration(fun)
}
\description{
To automatically assign argument objects with the default value, see \code{\link{declareArgs}}
}
\seealso{
Other make functions: 
\code{\link{makeArgs}()},
\code{\link{makeImports}()}
}
\concept{make functions}
