% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeArgs}
\alias{makeArgs}
\title{Make Args}
\usage{
makeArgs(fun)
}
\description{
Make Args
}
\seealso{
Other make functions: 
\code{\link{makeArgDeclaration}()},
\code{\link{makeDescURL}()},
\code{\link{makeImports}()}
}
\concept{make functions}
