% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeDefaultArgs}
\alias{makeDefaultArgs}
\title{Make a Formal Argument Skeleton with Default Values}
\usage{
makeDefaultArgs(fun)
}
\arguments{
\item{fun}{Function object.}
}
\description{
Retrieve the formal arguments with the default values to copy-and-paste as formal arguments of another function. To copy-and-paste formal arguments as part of an internal function, see \code{\link{makeInternalArgs}}.
}
\examples{
\dontrun{
if (interactive()) {
  makeDefaultArgs(read.csv)
}
}

}
