% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeDescription}
\alias{makeDescription}
\title{Make URLS For DESCRIPTION}
\usage{
makeDescription(path = getwd(), github_user, repo)
}
\description{
Make URLS For DESCRIPTION
}
\seealso{
Other make functions: 
\code{\link{makeArgDeclaration}()},
\code{\link{makeArgs}()},
\code{\link{makeDescURL}()},
\code{\link{makeImports}()}

Other DESCRIPTION functions: 
\code{\link{makeDescURL}()},
\code{\link{makeImports}()},
\code{\link{read_description}()}
}
\concept{DESCRIPTION functions}
\concept{make functions}
