% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeImports}
\alias{makeImports}
\title{Make DESCRIPTION Imports}
\usage{
makeImports()
}
\description{
Make DESCRIPTION Imports
}
\seealso{
Other make functions: 
\code{\link{makeArgDeclaration}()},
\code{\link{makeArgs}()},
\code{\link{makeDescriptionLinks}()},
\code{\link{makeDescription}()}

Other DESCRIPTION functions: 
\code{\link{makeDescriptionLinks}()},
\code{\link{makeDescription}()},
\code{\link{read_description}()}
}
\concept{DESCRIPTION functions}
\concept{make functions}
