% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeInternalArgs}
\alias{makeInternalArgs}
\title{Make a Formal Argument Skeleton with Default Values}
\usage{
makeInternalArgs(fun)
}
\arguments{
\item{fun}{PARAM_DESCRIPTION}
}
\description{
Retrieve the formal arguments assigned to itself  to copy-and-paste as an internal function call. To copy-and-paste formal arguments with their default values, see \code{\link{makeDefaultArgs}}.
}
\examples{
\dontrun{
if(interactive()){
       makeInternalArgs(read.csv)
 }
}

}
