% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readme.R
\name{makeReadmeInstallation}
\alias{makeReadmeInstallation}
\title{Make Installation Section in README}
\usage{
makeReadmeInstallation(pkg = basename(getwd()))
}
\description{
Make Installation Section in README
}
