% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{map_colors}
\alias{map_colors}
\title{FUNCTION_TITLE}
\usage{
map_colors(x, color_assignment)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{color_assignment}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
