% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_colors.R
\name{map_to_color}
\alias{map_to_color}
\title{Map Vector to Colors}
\usage{
map_to_color(x, color_assignment, other)
}
\arguments{
\item{x}{Character vector.}

\item{color_assignment}{Vector of new values named with \code{x} values. NA values
are viewed as "NA" strings, meaning that NA and "NA" will
be treated as the same value. Map assignment should therefore be 'c("NA" = "{new_value}")'.}

\item{other}{The value to map all other values not in \code{color_assignment}. Required only if
\code{color_assignment} is incomplete.}
}
\value{
A character vector of the same length as x.
}
\description{
Will abort if \code{color_assignment} does not contain valid colors in \code{colors_hex_map()} or \code{hex_colors_map()}.
}
\details{
This function converts \code{x} to factor, recodes the factor, and then
returns the recoded factor vector as character using \code{map_to_value()}.

DETAILS
}
