% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_cache.R
\name{preknit_cache}
\alias{preknit_cache}
\title{Cache or Load an Object Depending on
Interactive Status}
\usage{
preknit_cache(object, dirs)
}
\arguments{
\item{object}{Object to be cached.}

\item{dirs}{Required. Path to the subdirectory within R.cache root path to
save the cache to.}
}
\value{
\preformatted{  If called in a non-interactive session, the object will be returned
}

in the parent frame without being declared.
}
\description{
If called in an interactive session, the object will be cached to be
subsequently loaded in a non-interactive session, such as when R markdown
is knit. A previous run of a given chunk is required and an error will be
thrown if the cached object does not exist.
}
\details{
This function is useful when reporting on a process that has already been
executed and does not require re-execution during the knitting process.
For example, a Neo4j database can be written in an interactive session at
which point the metadata for reporting purposes can be logged. Instead of forcing
the database to be instantiated when the R markdown is knit, knitr will
only pull up the cached log.
}
\seealso{
\code{\link[R.cache]{saveCache}},\code{\link[R.cache]{findCache}},\code{\link[R.cache]{loadCache}}
}
