% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_tree.R
\name{print_file_tree}
\alias{print_file_tree}
\title{Print File Tree}
\usage{
print_file_tree(path, root_name = basename(path))
}
\description{
Print File Tree
}
\seealso{
\code{\link[tibble]{tibble}}
\code{\link[dplyr]{mutate}}
\code{\link[glue]{glue}}
\code{\link[data.tree]{as.Node}},\code{\link[data.tree]{as.Node.data.frame}},\code{\link[data.tree]{as.Node.dendrogram}},\code{\link[data.tree]{as.Node.list}},\code{\link[data.tree]{as.Node.phylo}},\code{\link[data.tree]{as.Node.rpart}}
}
