% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_table.R
\name{print_list}
\alias{print_list}
\title{Print Attribute Value Pairs as a Table}
\usage{
print_list(
  ...,
  exclude_colnames = TRUE,
  align = c("l", "l"),
  full_width = FALSE,
  position = "left",
  htmltable_class = "lightable-minimal"
)
}
\arguments{
\item{...}{Named list to print.}

\item{exclude_colnames}{Include "Attribute" and "Value" names?, Default: TRUE}

\item{align}{Column alignment: a character vector consisting of \code{'l'}
(left), \code{'c'} (center) and/or \code{'r'} (right). By default or if
\code{align = NULL}, numeric columns are right-aligned, and other columns
are left-aligned. If \code{length(align) == 1L}, the string will be
expanded to a vector of individual letters, e.g. \code{'clc'} becomes
\code{c('c', 'l', 'c')}, unless the output format is LaTeX.}

\item{full_width}{A \code{TRUE} or \code{FALSE} variable controlling whether the HTML
table should have 100\\% width. Since HTML and pdf have different flavors on
the preferable format for \code{full_width}. If not specified, a HTML table will
have full width by default but this option will be set to \code{FALSE} for a
LaTeX table}

\item{position}{This is the "real" or say floating position for the latex
table environment. The \code{kable} only puts tables in a table environment when
a caption is provided. That is also the reason why your tables will be
floating around if you specify captions for your table. Possible choices are
\code{h} (here), \code{t} (top, default), \code{b} (bottom) and \code{p} (on a dedicated page).}

\item{htmltable_class}{Options to use the in-house lightable themes.
Choices include \code{lightable-minimal}, \code{lightable-classic},
\code{lightable-classic-2}, \code{lightable-material}, \code{lightable-striped} and
\code{lightable-hover}. If you have your customized style sheet loaded which
defines your own table class, you can also load it here.}
}
\description{
Print user-provided attribute-value pairs in
a clean table format.
}
\seealso{
\code{\link[rlang]{list2}},\code{\link[rlang]{parse_expr}}
\code{\link[kableExtra]{kbl}},\code{\link[kableExtra]{kable_styling}}
}
