% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmail_filters.R
\name{process_gmail_filters_xml}
\alias{process_gmail_filters_xml}
\title{Process Gmail Filters XML}
\usage{
process_gmail_filters_xml(xml_path)
}
\description{
Process Gmail Filters XML
}
\seealso{
Other Google: 
\code{\link{authorize_google}()},
\code{\link{get_gmail_inbox}()},
\code{\link{get_gmail_labels}()},
\code{\link{get_gmail_messages_by_label}()},
\code{\link{get_unread_gmail}()},
\code{\link{get_unread_unlabelled_emails}()}
}
\concept{Google}
