% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_package.R
\name{read_description}
\alias{read_description}
\title{Read DESCRIPTION as a Dataframe}
\usage{
read_description(path = getwd())
}
\description{
Read the DESCRIPTION file as a 2-column dataframe for the header name along with the value/s following the colon of that header. To make these values in the R console, see \code{\link{makeDescription}}.
}
\examples{
read_description()

}
\seealso{
\code{\link[stringr]{str_replace}}
\code{\link[tibble]{tibble}}
\code{\link[tidyr]{unite}}

Other DESCRIPTION functions: 
\code{\link{makeDescURL}()},
\code{\link{makeDescription}()},
\code{\link{makeImports}()}
}
\concept{DESCRIPTION functions}
