% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_package.R
\name{read_namespace}
\alias{read_namespace}
\title{Read NAMESPACE as a List}
\usage{
read_namespace(path = getwd())
}
\description{
Read the NAMESPACE file as a list split by whether it is tagged as an export, import, or an importFrom. The export and imports are returned as a list of package names while the importFrom is a dataframe of the package name and the function from the package.
}
\examples{
read_namespace()
}
\seealso{
Other R Package functions: 
\code{\link{read_description}()}
}
\concept{NAMESPACE functions}
\concept{R Package functions}
