% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{remove_report}
\alias{remove_report}
\title{Remove a Report}
\usage{
remove_report(issue_key, report_title)
}
\arguments{
\item{issue_key}{Issue identifier.}

\item{report_title}{Name of the new report. It cannot contain any
forward slashes.}
}
\description{
Removes all the directories and files based on a issue_key and report_title pair.
No changes are made in cases where the provided pair does not exist in the
working directory or the function is not run interactively.
}
\seealso{
\code{\link[cli]{cat_line}},\code{\link[cli]{cli_alert}},\code{\link[cli]{cli_bullets}}
\code{\link[secretary]{press_enter}}
}
