% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_new_package.R
\name{setup_package_docs}
\alias{setup_package_docs}
\title{Setup Package Documentation}
\usage{
setup_package_docs(open = FALSE, spellcheck_vignettes = TRUE)
}
\arguments{
\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}

\item{spellcheck_vignettes}{Spell check is automatically
enabled for the new package, but if TRUE, will also
spellcheck all rmd and rnw files in the vignettes/ folder.}
}
\description{
Adds the package doc \verb{R/<package>-package.R},
\code{README.Rmd}, Code of Conduct, and \code{NEWS.md} documents.
Spellcheck is also added usig this function.
}
