% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_project.R
\name{start_pkg_proj}
\alias{start_pkg_proj}
\title{Start a Package Project}
\usage{
start_pkg_proj(path, open = FALSE, spellcheck_vignettes = TRUE)
}
\arguments{
\item{path}{A path. If it exists, it is used. If it does not exist, it is
created, provided that the parent path exists.}

\item{open}{Open the newly created file for editing? Happens in RStudio, if
applicable, or via \code{\link[utils:file.edit]{utils::file.edit()}} otherwise.}

\item{spellcheck_vignettes}{Spell check is automatically enabled for the new package, but if TRUE, will also spell check all rmd and rnw files in the vignettes/ folder.}
}
\description{
Similar suite of functions as \code{\link{make_new_package}} that can occur every time a package is started (i.e. not a one-time setup function.)
}
\seealso{
\code{\link[usethis]{use_package_doc}},\code{\link[usethis]{use_readme_rmd}},\code{\link[usethis]{use_news_md}},\code{\link[usethis]{use_spell_check}},\code{\link[usethis]{git_vaccinate}}
\code{\link[utils]{install.packages}}
}
