% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color2.R
\name{unshaded_base_colors}
\alias{unshaded_base_colors}
\title{FUNCTION_TITLE}
\usage{
unshaded_base_colors()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}}
}
