% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_to_vctr.R
\name{vctr_to_str_nl}
\alias{vctr_to_str_nl}
\title{Convert a vector of values into a string}
\usage{
vctr_to_str_nl(vector, quote = "'")
}
\description{
Store values neatly in an Excel or other dataframe for summary or record-keeping, but is in a format that can be readily parsed back to an expression if need be. This is important in cases where the names of columns want to be saved for a dataframe in a particular script to reference back or to QA later on. The string can be converted back to a vector by its counterpart function string_to_vector() in this package.
}
\examples{

# String to Vector
test_vctr <- "c('A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J')"
str_to_vctr(test_vctr)

test_vctr <- 'c("A", "B", "C", "D", "E", "F", "G", "H", "I", "J")'
str_to_vctr(test_vctr)

test_vctr <- 'c(\"A\", \"B\", \"C\", \"D\", \"E\", \"F\", \"G\", \"H\", \"I\", \"J\")'
str_to_vctr(test_vctr)


# Vector to String
vctr_to_str(LETTERS[1:5])

vctr_to_str(LETTERS[1:5], quote = "\"")
}
