% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{view_color_shades}
\alias{view_color_shades}
\title{FUNCTION_TITLE}
\usage{
view_color_shades(...)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{list2}}
\code{\link[stringr]{str_remove}}
\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}}
}
