% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{view_color_shades_regex}
\alias{view_color_shades_regex}
\title{View Colors Via Regex With Shades}
\usage{
view_color_shades_regex(pattern, invert = FALSE)
}
\arguments{
\item{pattern}{character string containing a \link[base]{regular expression}
    (or character string for \code{fixed = TRUE}) to be matched
    in the given character vector.  Coerced by
    \code{\link[base]{as.character}} to a character string if possible.  If a
    character vector of length 2 or more is supplied, the first element
    is used with a warning.  Missing values are allowed except for
    \code{regexpr} and \code{gregexpr}.}

\item{invert}{logical.  If \code{TRUE} return indices or values for
    elements that do \emph{not} match.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
View Colors Via Regex With Shades
}
