% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{view_colors}
\alias{view_colors}
\title{View Colors in the Plots Pane}
\usage{
view_colors(..., labels = TRUE, borders = NULL, cex_label = 0.5, ncol = NULL)
}
\arguments{
\item{...}{Color names or hex codes.}

\item{labels}{If TRUE, the plot is labelled with the color name. Default: TRUE}

\item{borders}{Color for rectangle border(s) that is passed to \code{rect()}, Default: NULL}

\item{cex_label}{Numeric character expansion factor; multiplied by par("cex") yields the final character size. NULL and NA are equivalent to 1.0, Default: 0.5}

\item{ncol}{How wide the plot should be, Default: NULL}
}
\description{
View Colors in the Plots Pane
}
\seealso{
\code{\link[rlang]{list2}}
}
