% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{view_colors_regex}
\alias{view_colors_regex}
\title{View Colors Via Regex}
\usage{
view_colors_regex(
  pattern,
  invert = FALSE,
  labels = TRUE,
  borders = NULL,
  cex_label = 0.5,
  ncol = NULL
)
}
\arguments{
\item{pattern}{character string containing a \link[base]{regular expression}
    (or character string for \code{fixed = TRUE}) to be matched
    in the given character vector.  Coerced by
    \code{\link[base]{as.character}} to a character string if possible.  If a
    character vector of length 2 or more is supplied, the first element
    is used with a warning.  Missing values are allowed except for
    \code{regexpr} and \code{gregexpr}.}

\item{invert}{logical.  If \code{TRUE} return indices or values for
    elements that do \emph{not} match.}

\item{labels}{If TRUE, the plot is labelled with the color name. Default: TRUE}

\item{borders}{Color for rectangle border(s) that is passed to \code{rect()}, Default: NULL}

\item{cex_label}{Numeric character expansion factor; multiplied by par("cex") yields the final character size. NULL and NA are equivalent to 1.0, Default: 0.5}

\item{ncol}{How wide the plot should be, Default: NULL}
}
\description{
View Colors Via Regex
}
