% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{view_labelled_colors}
\alias{view_labelled_colors}
\title{FUNCTION_TITLE}
\usage{
view_labelled_colors(
  ...,
  labels = NULL,
  borders = NULL,
  cex_label = 0.5,
  ncol = NULL
)
}
\arguments{
\item{...}{PARAM_DESCRIPTION}

\item{labels}{PARAM_DESCRIPTION, Default: TRUE}

\item{borders}{PARAM_DESCRIPTION, Default: NULL}

\item{cex_label}{PARAM_DESCRIPTION, Default: 0.5}

\item{ncol}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[rlang]{list2}}
}
