% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pkg.R
\name{write_data_file}
\alias{write_data_file}
\title{Write Processing File}
\usage{
write_data_file(path)
}
\description{
To write the usethis.R file to \code{data-raw}, the data.R file to \code{R}, and run
\code{devtools::document()}, see \code{\link{document_data}}.
}
