% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pkg.R
\name{write_schema_to_csvs}
\alias{write_schema_to_csvs}
\title{Export a Schema to CSV Files}
\usage{
write_schema_to_csvs(
  conn,
  schema,
  ...,
  path = getwd(),
  na = "NA",
  append = FALSE,
  col_names = !append,
  quote_escape = "double",
  eol = "\\n"
)
}
\arguments{
\item{conn}{Connection to a Postgres database.}

\item{...}{Optional. Character vectors of names of tables to export. If missing,
all the tables in the given schema are exported.}

\item{path}{Path to the package root that contains the \verb{data-raw/}, \verb{data/}, and \verb{R/} folders. `}

\item{na}{String used for missing values. Defaults to NA. Missing values
will never be quoted; strings with the same value as \code{na} will
always be quoted.}

\item{append}{If \code{FALSE}, will overwrite existing file. If \code{TRUE},
will append to existing file. In both cases, if the file does not exist a new
file is created.}

\item{col_names}{If \code{FALSE}, column names will not be included at the top of the file. If \code{TRUE},
column names will be included. If not specified, \code{col_names} will take the opposite value given to \code{append}.}

\item{quote_escape}{\Sexpr[results=rd, stage=render]{lifecycle::badge("deprecated")}, use the \code{escape} argument instead.}

\item{eol}{The end of line character to use. Most commonly either \code{"\n"} for
Unix style newlines, or \code{"\r\n"} for Windows style newlines.}
}
\description{
Write an entire schema or a subset of tables within a schema to a csv for each
table to the given path. The csvs are named according to the source table name.
}
