% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{auto_map_by_color_group}
\alias{auto_map_by_color_group}
\title{Randomly Select Colors By Color Group}
\usage{
auto_map_by_color_group(x, color_group_assignment, seed = 100)
}
\arguments{
\item{x}{Vector to map.}

\item{color_group_assignment}{Vector named by value in X and
values as one of c('pink', 'red', 'orange', 'yellow', 'brown', 'green',
'cyan', 'blue', 'purple', 'white', 'gray').}
}
\description{
Randomly Select Colors By Color Group
}
\seealso{
\code{\link[dplyr]{filter}}
}
