% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color_to_hex}
\alias{color_to_hex}
\title{FUNCTION_TITLE}
\usage{
color_to_hex(x, alpha = FALSE)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{alpha}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[grDevices]{col2rgb}},\code{\link[grDevices]{rgb}}
}
