% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{colors_hex_map}
\alias{colors_hex_map}
\title{FUNCTION_TITLE}
\usage{
colors_hex_map()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
