% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{hex_colors_map}
\alias{hex_colors_map}
\title{Map Between Hex Codes and Base R Colors}
\usage{
hex_colors_map(x)
}
\value{
Vector of color names named by their corresponding hexadecimal code.
}
\description{
The inverse map of all colors returned by base::colors() are
mapped to their corresponding hex codes.
}
