% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeDescURL}
\alias{makeDescURL}
\title{Make URLS For DESCRIPTION}
\usage{
makeDescURL(github_user, repo)
}
\description{
Make URLS For DESCRIPTION
}
\seealso{
Other make functions: 
\code{\link{makeArgDeclaration}()},
\code{\link{makeArgs}()},
\code{\link{makeImports}()}

Other DESCRIPTION functions: 
\code{\link{makeImports}()}
}
\concept{DESCRIPTION functions}
\concept{make functions}
