% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{makeImports}
\alias{makeImports}
\title{Make Arg Declaration}
\usage{
makeImports()
}
\description{
To automatically assign argument objects with the default value, see \code{\link{declareArgs}}
}
\seealso{
Other make functions: 
\code{\link{makeArgDeclaration}()},
\code{\link{makeArgs}()}
}
\concept{DESCRIPTION functions}
\concept{make functions}
