% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_colors.R
\name{map_colors}
\alias{map_colors}
\title{Map Vector to Colors}
\usage{
map_colors(x, color_assignment, other)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{color_assignment}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
Colors can be listed using \code{colors_hex_map()}.
}
\details{
DETAILS
}
