% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_value.R
\name{map_to_value}
\alias{map_to_value}
\title{Map Character Vector to New Values}
\usage{
map_to_value(x, map_assignment, na, other)
}
\arguments{
\item{x}{Character vector.}

\item{other}{The value to map all other values not in \code{map_assignment}.  Required only if
\code{map_assignment} is incomplete.}
}
\value{
A character vector of the same length as x.
}
\description{
Map Character Vector to New Values
}
\details{
This function converts \code{x} to factor, recodes the factor, and then
returns the recoded factor vector as character.
}
\examples{
x <- c('T', 'R', 'H', 'W', 'P', 'NA', NA_character)
map_to_value(
  x = x,
  map_assignment =
  c('T' = 'red',
    'R' = 'green'),
  other = 'black'
)

map_to_value(
  x = x,
  map_assignment =
  c('T' = 'red',
    'R' = 'green',
    'NA' = 'white'),
  other = 'black'
)
}
\concept{mapping functions}
