% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_to_value.R
\name{map_to_value}
\alias{map_to_value}
\title{FUNCTION_TITLE}
\usage{
map_to_value(x, map_assignment, other)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{map_assignment}{Named vector that is the inverse of what is expected in the forcats fct_recode arguments.
Map assignments to \code{NA_character_} should use \code{NA = {value}}.}

\item{other}{The value to map all other values not in the \code{map_assignment}.}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
