% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pkg.R
\name{prep_data_repo_dirs}
\alias{prep_data_repo_dirs}
\title{Create Repo Directories}
\usage{
prep_data_repo_dirs(path = getwd())
}
\description{
Creates the \code{data-raw}, \code{data}, and \code{R} directories within the \code{path} if they do not already exist.
}
