% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_new_package.R
\name{setup_tidy_git}
\alias{setup_tidy_git}
\title{Setup Tidy Git}
\usage{
setup_tidy_git()
}
\description{
Setup both GitHub and tidy style settings in the usethis package.
}
\seealso{
\code{\link[usethis]{use_git}},\code{\link[usethis]{git_vaccinate}},\code{\link[usethis]{use_github}},\code{\link[usethis]{use_pipe}},\code{\link[usethis]{use_tibble}},\code{\link[usethis]{use_tidy_github_actions}}
}
