% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-start_project.R, R/start_project.R
\name{start_pkg_proj}
\alias{start_pkg_proj}
\title{FUNCTION_TITLE}
\usage{
start_pkg_proj(path, open = FALSE, spellcheck_vignettes = TRUE)

start_pkg_proj(path, open = FALSE, spellcheck_vignettes = TRUE)
}
\arguments{
\item{path}{PARAM_DESCRIPTION}

\item{open}{PARAM_DESCRIPTION, Default: FALSE}

\item{spellcheck_vignettes}{PARAM_DESCRIPTION, Default: TRUE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION

Similar suite of functions as \code{\link{make_new_package}} that can occur every time a package is started (i.e. not a one-time setup function.)
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[usethis]{use_package_doc}},\code{\link[usethis]{use_readme_rmd}},\code{\link[usethis]{use_news_md}},\code{\link[usethis]{use_spell_check}},\code{\link[usethis]{git_vaccinate}}
\code{\link[utils]{install.packages}}

\code{\link[usethis]{use_package_doc}},\code{\link[usethis]{use_readme_rmd}},\code{\link[usethis]{use_news_md}},\code{\link[usethis]{use_spell_check}},\code{\link[usethis]{git_vaccinate}}
\code{\link[utils]{install.packages}}
}
