% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{view_color_shades}
\alias{view_color_shades}
\title{View Each Color and Its Shades}
\usage{
view_color_shades(
  ...,
  labels = TRUE,
  borders = NULL,
  cex_label = 0.5,
  ncol = NULL
)
}
\arguments{
\item{...}{Base names of colors from \code{base_colors()}}

\item{labels}{If TRUE, the plot is labelled with the color name. Default: TRUE}

\item{borders}{Color for rectangle border(s) that is passed to \code{rect()}, Default: NULL}

\item{cex_label}{Numeric character expansion factor; multiplied by par("cex") yields the final character size. NULL and NA are equivalent to 1.0, Default: 0.5}

\item{ncol}{How wide the plot should be, Default: NULL}
}
\description{
View Each Color and Its Shades
}
\seealso{
\code{\link[rlang]{list2}}
\code{\link[stringr]{str_remove}}
\code{\link[dplyr]{filter}},\code{\link[dplyr]{select}}
}
