% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color2.R
\name{view_color_shades_regex}
\alias{view_color_shades_regex}
\title{FUNCTION_TITLE}
\usage{
view_color_shades_regex(pattern, invert = FALSE)
}
\arguments{
\item{pattern}{PARAM_DESCRIPTION}

\item{invert}{PARAM_DESCRIPTION, Default: FALSE}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
