% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color2.R
\name{view_colors_regex}
\alias{view_colors_regex}
\title{FUNCTION_TITLE}
\usage{
view_colors_regex(
  pattern,
  invert = FALSE,
  labels = TRUE,
  borders = NULL,
  cex_label = 0.5,
  ncol = NULL
)
}
\arguments{
\item{pattern}{PARAM_DESCRIPTION}

\item{invert}{PARAM_DESCRIPTION, Default: FALSE}

\item{labels}{PARAM_DESCRIPTION, Default: TRUE}

\item{borders}{PARAM_DESCRIPTION, Default: NULL}

\item{cex_label}{PARAM_DESCRIPTION, Default: 0.5}

\item{ncol}{PARAM_DESCRIPTION, Default: NULL}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
